Ensure bcs cache key is consistent in parallel
Construct bc nodes currently requires a parallel communication
round (and is hence collective).    Ensure that we maintain the
correct ordering when running in parallel, by sorting the bcs to
access on a deterministic basis.  Fixes #1212.

A better mechanism would be to remove the requirement that
constructing boundary nodes requires synchronisation.  This would be
more work, however.